Minor package consolidation
