+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/wascore/common/RedirectPolicy.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.wascore.authentication;
+package com.microsoft.applicationinsights.agent.internal.wascore.common;
 
 import com.azure.core.http.HttpPipelineCallContext;
 import com.azure.core.http.HttpPipelineNextPolicy;
