+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/wascore/perfcounter/JmxMetricPerformanceCounter.java
@@ -24,7 +24,7 @@ package com.microsoft.applicationinsights.agent.internal.wascore.perfcounter;
 import com.azure.monitor.opentelemetry.exporter.implementation.models.TelemetryItem;
 import com.microsoft.applicationinsights.agent.internal.wascore.TelemetryClient;
 import com.microsoft.applicationinsights.agent.internal.wascore.TelemetryUtil;
-import com.microsoft.applicationinsights.agent.internal.wascore.jmx.JmxAttributeData;
+import com.microsoft.applicationinsights.agent.internal.wascore.perfcounter.jmx.JmxAttributeData;
 import java.util.Collection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
