+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/wascore/perfcounter/jmx/JmxAttributeData.java
@@ -19,15 +19,18 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.wascore.config.connection;
+package com.microsoft.applicationinsights.agent.internal.wascore.perfcounter.jmx;
 
-public class InvalidConnectionStringException extends Exception {
+/**
+ * Represents JMX data of an Attribute The display name The name of the attribute The type of the
+ * attribute.
+ */
+public final class JmxAttributeData {
+  public final String metricName;
+  public final String attribute;
 
-  InvalidConnectionStringException(String message) {
-    super(message);
-  }
-
-  InvalidConnectionStringException(String message, Throwable cause) {
-    super(message, cause);
+  public JmxAttributeData(String metricName, String attribute) {
+    this.attribute = attribute;
+    this.metricName = metricName;
   }
 }
