+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/wascore/perfcounter/jmx/JmxDataFetcher.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.wascore.jmx;
+package com.microsoft.applicationinsights.agent.internal.wascore.perfcounter.jmx;
 
 import java.lang.management.ManagementFactory;
 import java.util.ArrayList;
