+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/wascore/perfcounter/jmx/JmxDataFetcherTest.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.wascore.jmx;
+package com.microsoft.applicationinsights.agent.internal.wascore.perfcounter.jmx;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
